Fixing IPN hash (part 2)
Hey, you may remember a while back I opened an issue for this and you have implemented it (thanks!). Well recently the sites which I use g2a on have had a lot of orders over 1000 and the IPNs have been failing. After debugging I found the cause of the issue was the hash - as number_format(int, 2) when over 1000 will include a comma and ends up breaking it. 

tl;dr - removes the thousands separator so it will no longer break for orders over 1000